Parse without use of regex to work around dependency conflict
The dependencies of the `regex` crate (especially `memchr`) clash
with my runtime dependencies which need to be `no_std`. I don't think
there's a general way to prevent feature clashes between build-time and
run-time dependencies yet.

So the most straightforward workaround seems to be to write the parsing
in this crate without relying on a regex.